Remove admin routes
