--- express-api/src/controllers/admin/accessRequests/IAccessRequestList.ts
@@ -1,8 +1,0 @@
-import { IAccessRequest } from '@/controllers/admin/accessRequests/IAccessRequest';
-
-export interface IAccessRequestList {
-  items: IAccessRequest[];
-  page: number;
-  quantity: number;
-  total: number;
-}
