--- express-api/src/controllers/admin/claims/IClaim.ts
@@ -1,10 +1,0 @@
-import { UUID } from 'crypto';
-import { IBaseEntity } from '@/controllers/common/IBaseEntity';
-
-export interface IClaim extends IBaseEntity {
-  id?: UUID;
-  name: string;
-  keycloakRoleId: UUID;
-  description: string;
-  isDisabled: boolean;
-}
