+++ express-api/src/express.ts
@@ -10,7 +10,6 @@ import middleware from '@/middleware';
 import constants from '@/constants';
 import { KEYCLOAK_OPTIONS } from '@/middleware/keycloak/keycloakOptions';
 import swaggerUi from 'swagger-ui-express';
-import { Roles } from '@/constants/roles';
 import swaggerJSON from '@/swagger/swagger-output.json';
 import errorHandler from '@/middleware/errorHandler';
 import { EndpointNotFound404 } from '@/constants/errors';
@@ -68,11 +67,11 @@ app.use(`/v2/health`, router.healthRouter);
 
 // Protected Routes
 app.use(`/v2/ltsa`, protectedRoute(), router.ltsaRouter);
-app.use(`/v2/admin`, protectedRoute([Roles.ADMIN]), router.adminRouter);
 app.use(`/v2/administrativeAreas`, protectedRoute(), router.administrativeAreasRouter);
 app.use(`/v2/agencies`, protectedRoute(), router.agenciesRouter);
 app.use('/v2/lookup', protectedRoute(), router.lookupRouter);
 app.use(`/v2/users`, protectedRoute(), router.usersRouter);
+app.use(`/v2/roles`, protectedRoute(), router.rolesRouter);
 app.use(`/v2/properties`, protectedRoute(), router.propertiesRouter);
 app.use(`/v2/parcels`, protectedRoute(), router.parcelsRouter);
 app.use(`/v2/buildings`, protectedRoute(), router.buildingsRouter);
