+++ express-api/src/routes/index.ts
@@ -3,7 +3,6 @@ import healthRouter from '@/routes/healthRouter';
 import buildingsRouter from '@/routes/buildingsRouter';
 import parcelsRouter from '@/routes/parcelsRouter';
 import lookupRouter from '@/routes/lookupRouter';
-import adminRouter from '@/routes/adminRouter';
 import usersRouter from '@/routes/usersRouter';
 import propertiesRouter from '@/routes/propertiesRouter';
 import projectsRouter from '@/routes/projectsRouter';
@@ -12,6 +11,7 @@ import reportsRouter from '@/routes/reportsRouter';
 import toolsRouter from '@/routes/toolsRouter';
 import agenciesRouter from '@/routes/agenciesRouter';
 import administrativeAreasRouter from '@/routes/administrativeAreasRouter';
+import rolesRouter from '@/routes/rolesRouter';
 
 const router = {
   administrativeAreasRouter,
@@ -20,7 +20,6 @@ const router = {
   buildingsRouter,
   parcelsRouter,
   lookupRouter,
-  adminRouter,
   usersRouter,
   propertiesRouter,
   projectsRouter,
@@ -28,6 +27,7 @@ const router = {
   reportsRouter,
   toolsRouter,
   agenciesRouter,
+  rolesRouter,
 };
 
 export default router;
