+++ express-api/tests/integration/admin/roles/roles.test.ts
@@ -2,7 +2,7 @@ import supertest from 'supertest';
 import { faker } from '@faker-js/faker';
 import app from '@/express';
 import { UUID } from 'crypto';
-import { IRole } from '@/controllers/admin/roles/IRole';
+import { IRole } from '@/controllers/roles/IRole';
 
 const request = supertest(app);
