+++ express-api/tests/integration/admin/users/users.test.ts
@@ -2,7 +2,7 @@ import supertest from 'supertest';
 import { faker } from '@faker-js/faker';
 import app from '@/express';
 import { UUID } from 'crypto';
-import { IUser } from '@/controllers/admin/users/IUser';
+import { IUser } from '@/controllers/users/IUser';
 
 const request = supertest(app);
